add HTTPServer structure
