--- river-monitoring-service/MQTTServer/src/main/java/org/mqttserver/policy/README.md
@@ -1,1 +1,0 @@
-this folder contains all policy in order to retrieve the correct response.
