--- river-monitoring-service/MQTTServer/src/main/java/org/mqttserver/service/Broker.java
@@ -1,10 +1,0 @@
-package org.mqttserver.service;
-
-import io.vertx.mqtt.MqttServer;
-
-public interface Broker {
-    boolean initialize();
-
-
-
-}
