+++ river-monitoring-service/Server/src/main/java/org/mqttserver/Main.java
@@ -1,0 +1,12 @@
+package org.mqttserver;
+
+import org.mqttserver.services.MQTT.Broker;
+import org.mqttserver.services.MQTT.BrokerImpl;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Welcome in Smart River Monitoring Server...");
+        Broker broker = new BrokerImpl();
+        broker.initialize(broker.getMqttServer());
+    }
+}
