+++ river-monitoring-service/Server/src/main/java/org/mqttserver/policy/README.md
@@ -1,0 +1,1 @@
+this folder contains all policies in order to retrieve the correct response.
