+++ river-monitoring-service/Server/src/main/java/org/mqttserver/serial/Reader.java
@@ -1,0 +1,6 @@
+package org.mqttserver.serial;
+
+public interface Reader {
+
+    String readDataFromSerial();
+}
