+++ river-monitoring-service/Server/src/main/java/org/mqttserver/serial/Writer.java
@@ -1,0 +1,6 @@
+package org.mqttserver.serial;
+
+public interface Writer {
+
+    void writeDataOnSerial(String message);
+}
