+++ river-monitoring-service/Server/src/main/java/org/mqttserver/services/HTTP/HTTPServer.java
@@ -1,0 +1,2 @@
+package org.mqttserver.services.HTTP;public interface HTTPServer {
+}
