+++ river-monitoring-service/Server/src/main/java/org/mqttserver/services/MQTT/Broker.java
@@ -1,0 +1,9 @@
+package org.mqttserver.services.MQTT;
+
+import io.vertx.mqtt.MqttServer;
+
+public interface Broker {
+    boolean initialize(MqttServer mqttServer);
+    MqttServer getMqttServer();
+
+}
