+++ river-monitoring-service/Server/src/main/java/org/mqttserver/services/MQTT/publisher/MQTTPublisher.java
@@ -1,0 +1,9 @@
+package org.mqttserver.services.MQTT.publisher;
+
+import io.vertx.mqtt.MqttException;
+
+public interface MQTTPublisher {
+    void publish(String topic, String message);
+
+
+}
