+++ river-monitoring-service/Server/src/main/java/org/mqttserver/services/MQTT/publisher/MQTTPublisherImpl.java
@@ -1,0 +1,11 @@
+package org.mqttserver.services.MQTT.publisher;
+
+import org.mqttserver.services.MQTT.BrokerImpl;
+
+public class MQTTPublisherImpl extends BrokerImpl {
+
+
+    public MQTTPublisherImpl() {
+        //TODO:
+    }
+}
