+++ river-monitoring-service/Server/src/main/java/org/mqttserver/services/MQTT/subscriber/MQTTSubscriber.java
@@ -1,0 +1,8 @@
+package org.mqttserver.services.MQTT.subscriber;
+
+import io.vertx.core.Handler;
+import io.vertx.mqtt.MqttException;
+
+public interface MQTTSubscriber {
+    void subscribe(String topic, Handler<String> messageHandler);
+}
