+++ river-monitoring-service/Server/src/main/java/org/mqttserver/services/MQTT/subscriber/MQTTSubscriberImpl.java
@@ -1,0 +1,12 @@
+package org.mqttserver.services.MQTT.subscriber;
+
+
+import org.mqttserver.services.MQTT.BrokerImpl;
+
+public class MQTTSubscriberImpl extends BrokerImpl {
+
+    public MQTTSubscriberImpl() {
+        //TODO
+    }
+
+}
